Fetching Services the Lazy Way
One more thing! We can actually do all of this with less code. In the controller, replace {dollar_sign}this->container->get() with just {dollar_sign}this->get()

This does the same thing as before.

Ok: here's the big, giant important takeaway so far: why do we add bundles to our app? Because bundles put more services in our container. And services are tools.

So my question now is: how can we configure these services so that they do exactly what we need them to?