I've got a pretty important new challenge. We're going to be rendering a lot of markdown... and we don't want to do this on every request - it's just too slow. We really need a way to cache the parsed markdown.
Hmm, so caching is yet another tool that we need. If we had a service that was really good at caching a string and letting us fetch it out later, that would be perfect! Fortunately, Symfony comes with a bundle called DoctrineCacheBundle that can give us exactly this.

Enabling DoctrineCacheBundle

First, double-check that you have the bundle in your composer.json file:

 65 lines  composer.json
{
... lines 2 - 17
  'require': {
... lines 19 - 22
   'doctrine/doctrine-cache-bundle': '^1.2',
... lines 24 - 62
  }
}
If for some reason you don't, use Composer to download it.

The bundle lives in the vendor/ directory, but it isn't enabled. Do that in the AppKernel class with new DoctrineCacheBundle()

Once you've enabled a bundle, there are usually two more steps: configure it, then use it. And of course, this has its own documentation that'll explain all of this. But guys, we're already getting really good at Symfony. I bet we can figure out how to use it entirely on our own.