Configuring a Cache Service
Now, remember the goal: to get a cache service we can use to avoid processing markdown on each request. When we added KnpMarkdownBundle, we magically had a new service. But with this bundle, we need to configure each service we want.

Open up config.yml and add doctrine_cache. Below that, add a providers key.

Next, the config dump has a name key. This Prototype comment above that is a confusing term that means that we can call this name anything we want. Let's make it my_markdown_cache.

You'll see how that's important in a second.

Finally, tell Doctrine what type of cache this is by setting type to file_system.

This is just one of the built-in types this bundle offers: its docs would tell you the others.

And that's it! In the terminal run ./bin/console debug:container and search for markdown_cache:

./bin/console debug:container markdown_cache
Et voila! We have a new service called doctrine_cache.providers.my_markdown_cache. That's the whole point of this bundle: we describe a cache system we want, and it configures a service for that. Now we're dangerous.