Experimenting with config_dev.yml
Let's play around a bit with the dev environment! Under monolog, uncomment the firephp line

This is a cool handler that will show you log messages right in your browser.

Head over to it and run 'Inspect Element'. Make sure that the URL will access the dev environment and then refresh. And check this out: a bunch of messages telling us what route was matched. Heck, we can even see what route was matched for our ajax call. To see this working, you'll need a FirePHP extension installed in your browser. In your app, Monolog is attaching these messages to your response headers, and the extension is reading those. We don't want this to happen on production, so we only enabled this in the dev environment.

Environments are awesome! So how could we use them only cache our markdown string in the prod environment?