We absolutely need to cache our markdown processing. But what if we need to tweak how the markdown renders? In that case, we don't want caching. Could we somehow disable caching in the dev environment only?
Yes! Copy the doctrine_cache from config.yml and paste it into config_dev.yml. Next, change type from file_system to array

The array type is basically a 'fake' cache: it won't ever store anything.

Yea, that's it! Head back to the browser and refresh. This is definitely not caching: it takes an entire second because of the sleep(). Try again. Still not caching! Now, change to the prod environment and refresh. Beautiful, this is still really, really fast. Well that was easy.