Creating a new Parameter
We can leverage parameters to do something really cool with our cache setup.

In the prod environment, we use the file_system cache. In dev, we use array. We can improve this. Create a new parameter called cache_type and set that to file_system. Scroll down and set type to %cache_type%.

Run over in the terminal to see if the parameter showed up:

./bin/console debug:container --parameters
It's right on top. Cool! Clear the cache in the prod environment so we can double-check everything is still working:

./bin/console cache:clear --env=prod
Ok good - now refresh using app.php. It's still loading fast - so we haven't broken anything... yet.