Why is parameters.yml Special?
I have a question: if config.yml imports parameters.yml, then why bother having this file at all? Why not just put all the parameters at the top of config.yml?

Here's why: parameters.yml holds any configuration that will be different from one machine where the code is deployed to another.

For example, your database password is most likely not the same as my database password and hopefully not the same as the production database password. But if we put that password right in the middle of config.yml, that would be a nightmare! In that scenario I would probably commit my password to git and then you would have to change it to your password but then try to not commit that change. Gross.

Instead of that confusing mess of seaweed, we use parameters in config.yml. This allows us to isolate all the machine-specific configuration to parameters.yml. And here's the final key: parameters.yml is not committed to the repository - you can see there's an entry for it in .gitignore:

 17 lines  .gitignore
/app/config/parameters.yml
... lines 2 - 17
Of course, if I just cloned this project, and I won't have a parameters.yml file: I have to create it manually. Actually, this is the exact reason for this other file: parameters.yml.dist:

20 lines  app/config/parameters.yml.dist
# This file is a template of what your parameters.yml file should look like
# Set parameters here that may be different on each deployment target of the app, e.g. development, staging, production.
# http://symfony.com/doc/current/best_practices/configuration.html#infrastructure-related-configuration
parameters:
  database_host:     127.0.0.1
  database_port:     ~
  database_name:     symfony
  database_user:     root
  database_password: ~
  # You should uncomment this if you want use pdo_sqlite
  # database_path: %kernel.root_dir%/data.db3
  mailer_transport:  smtp
  mailer_host:       127.0.0.1
  mailer_user:       ~
  mailer_password:   ~
  # A secret key that's used to generate certain security-related tokens
  secret:            ThisTokenIsNotSoSecretChangeIt
This is not read by Symfony, it's just a template of all of the parameters this project needs. If you add or remove things from the parameters.yml, be sure to add or remove them from parameters.yml.dist. You do commit this file to git.

 Tip
Due to a post-install command in your composer.json, after running composer install, Symfony will read parameters.yml.dist and ask you to fill in any values that are missing from parameters.yml.